Refactoring and fixing bugs
